+++ wakatime/main.py
@@ -30,9 +30,11 @@ sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), 'pac
 
 from .__about__ import __version__
 from .compat import u, open, is_py3
+from .constants import SUCCESS, API_ERROR, CONFIG_FILE_PARSE_ERROR
 from .logger import setup_logging
 from .offlinequeue import Queue
 from .packages import argparse
+from .packages import requests
 from .packages.requests.exceptions import RequestException
 from .project import get_project_info
 from .session_cache import SessionCache
@@ -290,7 +292,7 @@ def send_heartbeat(project=None, branch=None, hostname=None, stats={}, key=None,
     """
 
     if not api_url:
-        api_url = 'https://wakatime.com/api/v1/heartbeats'
+        api_url = 'https://api.wakatime.com/api/v1/heartbeats'
     if not timeout:
         timeout = 30
     log.debug('Sending heartbeat to api at %s' % api_url)
@@ -368,7 +370,7 @@ def send_heartbeat(project=None, branch=None, hostname=None, stats={}, key=None,
     else:
         response_code = response.status_code if response is not None else None
         response_content = response.text if response is not None else None
-        if response_code == 201:
+        if response_code == requests.codes.created or response_code == requests.codes.accepted:
             log.debug({
                 'response_code': response_code,
             })
@@ -408,7 +410,7 @@ def execute(argv=None):
 
     args, configs = parseArguments()
     if configs is None:
-        return 103 # config file parsing error
+        return CONFIG_FILE_PARSE_ERROR
 
     setup_logging(args, __version__)
 
@@ -418,7 +420,7 @@ def execute(argv=None):
             log.debug(u('Skipping because matches exclude pattern: {pattern}').format(
                 pattern=u(exclude),
             ))
-            return 0
+            return SUCCESS
 
         if args.entity_type != 'file' or os.path.isfile(args.entity):
 
@@ -464,12 +466,12 @@ def execute(argv=None):
                     )
                     if not sent:
                         break
-                return 0 # success
+                return SUCCESS
 
-            return 102 # api error
+            return API_ERROR
 
         else:
             log.debug('File does not exist; ignoring this heartbeat.')
-            return 0
+            return SUCCESS
     except:
         log.traceback()